[feat] rtk-thunk connect
