+++ .idea/workspace.xml
@@ -3,13 +3,9 @@
   <component name="ChangeListManager">
     <list default="true" id="cee230d0-8ba8-4605-8330-6eb290caf7f1" name="Default Changelist" comment="">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/public/img/filtericon.png" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/components/mbti/MbtiBody.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/mbti/MbtiBody.tsx" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/components/mbti/MbtiFooter.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/mbti/MbtiFooter.tsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/components/mbti/MbtiSelect.tsx" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/src/pages/_document.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/pages/_document.tsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/pages/profile/[id].tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/pages/profile/[id].tsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/tsconfig.json" beforeDir="false" afterPath="$PROJECT_DIR$/tsconfig.json" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/store/slices/mbti/mbtiSlice.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/store/slices/mbti/mbtiSlice.tsx" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -67,6 +63,10 @@
       <workItem from="1656994417321" duration="490000" />
       <workItem from="1657083531682" duration="75000" />
       <workItem from="1657160769954" duration="1000" />
+      <workItem from="1657677044368" duration="67000" />
+      <workItem from="1657677113299" duration="2144000" />
+      <workItem from="1657698961887" duration="419000" />
+      <workItem from="1657699385254" duration="2685000" />
     </task>
     <servers />
   </component>
