+++ src/pages/_document.tsx
@@ -44,7 +44,7 @@ class MyDocument extends Document {
             as="style"
             href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.4/dist/web/static/pretendard.css"
           />
-          {/* <script src="https://developers.kakao.com/sdk/js/kakao.js"></script> */}
+           <script async src="https://developers.kakao.com/sdk/js/kakao.js"></script>
           <link
             href="https://fonts.googleapis.com/css2?family=Prompt:wght@100;400;700&display=swap"
             rel="stylesheet"
